+++ Vault/Changes.md
@@ -15,6 +15,8 @@ Supports **Flow v2.4**
 ### New methods
 - **ManyModelAccess**
   - Added `.findColumnMap(...)` and `.findColumnMultiMap(...)`
+### Other changes
+- **View**`.exists(...)` now supports joins
 
 ## v1.18 - 22.01.2024
 This update focuses on improving grouped (i.e. multi-joined) result parsing.